Yuqiang/merge coco ozar round 1 (#7002)
* merge app/core

* merge app/lib

* coffee style

* fix merge bug

* fix merge bug

* remove leading /

* Update app/core/utils.coffee


* fix coco teachers/resource & teachers/license

---------
