+++ app/components/common/PassThrough.vue
@@ -1,0 +1,3 @@
+<template>
+  <router-view />
+</template>
