+++ app/core/social-handlers/GoogleClassroomHandler.js
@@ -2,6 +2,8 @@ import api from 'core/api'
 import CocoClass from 'core/CocoClass'
 const SCOPE = 'https://www.googleapis.com/auth/classroom.courses.readonly https://www.googleapis.com/auth/classroom.profile.emails https://www.googleapis.com/auth/classroom.rosters.readonly'
 
+const utils = require('core/utils')
+
 const GoogleClassroomAPIHandler = class GoogleClassroomAPIHandler extends CocoClass {
 
   constructor () {
@@ -32,7 +34,7 @@ const GoogleClassroomAPIHandler = class GoogleClassroomAPIHandler extends CocoCl
   loadStudentsFromAPI (googleClassroomId, nextPageToken='') {
     return new Promise((resolve, reject) => {
       const fun = () => gapi.client.load('classroom', 'v1', () => {
-        gapi.client.classroom.courses.students.list({access_token: application.gplusHandler.token(), courseId: googleClassroomId, pageToken: nextPageToken})
+        gapi.client.classroom.courses.students.list({ access_token: application.gplusHandler.token(), courseId: googleClassroomId, pageToken: nextPageToken })
         .then((r) => {
             resolve(r.result || {})
           })
@@ -64,7 +66,11 @@ module.exports = {
     try {
       let gClass = me.get('googleClassrooms').find((c)=>c.id==gcId)
       if (gClass) {
-        gClass.importedToOzaria = true
+        if (utils.isCodeCombat) {
+          gClass.importedToCoco = true
+        } else {
+          gClass.importedToOzaria = true
+        }
         await new Promise(me.save().then)
       }
       else {
@@ -85,19 +91,18 @@ module.exports = {
       const importedClassroomsNames = importedClassrooms.map((c) => {
         return { id: c.id, name: c.name }
       })
-
       const classrooms = me.get('googleClassrooms') || []
       let mergedClassrooms = []
       importedClassroomsNames.forEach((imported) => {
         const cl = classrooms.find((c) => c.id == imported.id)
-        mergedClassrooms.push({...cl, ...imported})
+        mergedClassrooms.push({ ...cl, ...imported })
       })
 
       // classrooms that were imported to coco/ozaria but no more exist in importedClassroomNames, i.e. have been removed from google classroom
       const mergedClassroomIds = mergedClassrooms.map((m) => m.id)
       const extraClassroomsImported = classrooms.filter((c) => (c.importedToCoco || c.importedToOzaria) && !(mergedClassroomIds.includes(c.id)))
       // set deletedFromGC, so that it gets filtered from the dropdown on the create classroom modal
-      // for example, a class that is importedToCoco but deleted from GC should not be available in the dropdown on ozaria
+      // for example, a class that is importedToOzaria but deleted from GC should not be available in the dropdown on coco
       extraClassroomsImported.forEach((e) => e.deletedFromGC = true)
       mergedClassrooms = mergedClassrooms.concat(extraClassroomsImported)
       me.set('googleClassrooms', mergedClassrooms)
@@ -140,9 +145,9 @@ module.exports = {
       const createdStudents = signupStudentsResult.filter((s) => !s.isError)
       const signupErrors = signupStudentsResult.filter((s) => s.isError && s.errorID != 'student-account-exists')
       const existingStudents = signupStudentsResult
-        .filter((s) => s.errorID == 'student-account-exists') // TODO update error id to 'account-exists' since it might contain teacher/individual accounts also
+        .filter((s) => s.errorID === 'student-account-exists') // TODO update error id to 'account-exists' since it might contain teacher/individual accounts also
         .map((s) => s.error) // error contains the user object here
-        .filter((s) => s.role === 'student') // filter only student accounts, discard existing individual/teacher accounts
+        .filter((s) => (utils.isCodeCombat || s.role === 'student')) // For Ozaria: filter only student accounts, discard existing individual/teacher accounts
 
       console.debug("Students created:", createdStudents)
 
@@ -154,9 +159,17 @@ module.exports = {
       const classroomNewMembers = createdStudents.concat(existingStudents.filter((s) => !cocoClassroom.members.includes(s._id)))
 
       if (classroomNewMembers.length > 0){
-        await store.dispatch('classrooms/addMembersToClassroom', { classroom: cocoClassroom, members: classroomNewMembers })
+        if (utils.isCodeCombat) {
+          await api.classrooms.addMembers({ classroomID: cocoClassroom._id, members: classroomNewMembers })
+        } else {
+          await store.dispatch('classrooms/addMembersToClassroom', { classroom: cocoClassroom, members: classroomNewMembers })
+        }
         noty ( {text: classroomNewMembers.length+' Students imported.', layout: 'topCenter', timeout: 3000, type: 'success' })
         return classroomNewMembers
+      }
+      else if (utils.isCodeCombat) {
+        console.error("No new students imported. Error:", signupStudentsResult)
+        return Promise.reject('No new students imported')
       }
       else if (signupErrors.length > 0) {
         console.error("No new students imported. Error:", signupErrors)
