--- app/lib/dynamicRequire.ozar.js
@@ -1,171 +1,0 @@
-module.exports = {
-  'views/AboutView': function () { return import(/* webpackChunkName: "AboutView" */ 'views/AboutView') },
-  'views/HomeView': function () { return import(/* webpackChunkName: "HomeView" */ 'views/HomeView') },
-  'views/HomeCNView': function () { return import(/* webpackChunkName: "HomeCNView" */ 'views/HomeCNView') },
-  'views/ContactCNView': function () { return import(/* webpackChunkName: "ContactCNView" */ 'views/ContactCNView') },
-  'views/ChinaBridgeView': function () { return import(/* webpackChunkName: "ContactGEEKView" */ 'views/ChinaBridgeView') },
-  'views/account/OAuthAIYouthView': function () { return import(/* webpackChunkName: "OAuthAIYouthView" */ 'views/account/OAuthAIYouthView') },
-  'views/account/MainAccountView': function () { return import(/* webpackChunkName: "account" */ 'views/account/MainAccountView') },
-  'views/account/AccountSettingsRootView': function () { return import(/* webpackChunkName: "AccountSettingsRoot" */ 'views/account/AccountSettingsRootView') },
-  'views/account/AccountSettingsView': function () { return import(/* webpackChunkName: "AccountSettings" */ 'views/account/AccountSettingsView') },
-  'views/account/UnsubscribeView': function () { return import(/* webpackChunkName: "account" */ 'views/account/UnsubscribeView') },
-  'views/account/PaymentsView': function () { return import(/* webpackChunkName: "account" */ 'views/account/PaymentsView') },
-  'views/account/SubscriptionView': function () { return import(/* webpackChunkName: "account" */ 'views/account/SubscriptionView') },
-  'views/account/InvoicesView': function () { return import(/* webpackChunkName: "account" */ 'views/account/InvoicesView') },
-  'views/account/PrepaidView': function () { return import(/* webpackChunkName: "account" */ 'views/account/PrepaidView') },
-  'views/admin/MainAdminView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/MainAdminView') },
-  'views/admin/CLAsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/CLAsView') },
-  'views/admin/AdminClassroomContentView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AdminClassroomContentView') },
-  'views/admin/AdminClassroomLevelsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AdminClassroomLevelsView') },
-  'views/admin/PartialUnitReleaseView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/PartialUnitReleaseView') },
-  'views/admin/AdminClassroomsProgressView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AdminClassroomsProgressView') },
-  'views/admin/FilesView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/FilesView') },
-  'views/admin/AnalyticsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AnalyticsView') },
-  'views/admin/AnalyticsSubscriptionsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AnalyticsSubscriptionsView') },
-  'views/admin/AdminLevelHintsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AdminLevelHintsView') },
-  'views/admin/AdminSubCancellationsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/AdminSubCancellationsView') },
-  'views/admin/SchoolCountsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/SchoolCountsView') },
-  'views/admin/SchoolLicensesView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/SchoolLicensesView') },
-  'views/admin/BaseView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/BaseView') },
-  'views/admin/DemoRequestsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/DemoRequestsView') },
-  'views/admin/TrialRequestsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/TrialRequestsView') },
-  'views/admin/UserCodeProblemsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/UserCodeProblemsView') },
-  'views/admin/PendingPatchesView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/PendingPatchesView') },
-  'views/admin/CodeLogsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/CodeLogsView') },
-  'views/admin/SkippedContactsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/SkippedContactsView') },
-  'views/admin/OutcomeReportResultView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/OutcomeReportResultView') },
-  'views/admin/OutcomesReportView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/OutcomesReportView') },
-  'views/admin/DesignElementsView': function () { return import(/* webpackChunkName: "admin" */ 'views/admin/DesignElementsView') },
-  'views/ai/AIView': function () { return import(/* webpackChunkName: "ai" */ 'views/ai/AIView') },
-  'views/teachers/DynamicAPCSPView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/DynamicAPCSPView') },
-  'views/minigames/ConditionalMinigameView': function() { return import(/* webpackChunkName: "artisans" */ 'views/minigames/ConditionalMinigameView') },
-  'views/artisans/ArtisansView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/ArtisansView') },
-  'views/artisans/LevelTasksView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/LevelTasksView') },
-  'views/artisans/SolutionProblemsView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/SolutionProblemsView') },
-  'views/artisans/LevelConceptMap': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/LevelConceptMap') },
-  'views/artisans/LevelGuidesView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/LevelGuidesView') },
-  'views/artisans/StudentSolutionsView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/StudentSolutionsView') },
-  'views/artisans/TagTestView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/TagTestView') },
-  'views/artisans/BulkLevelEditView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/BulkLevelEditView') },
-  'views/artisans/SandboxView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/SandboxView') },
-  'views/artisans/ThangTasksView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/ThangTasksView') },
-  'views/artisans/ArenaBalancerView': function () { return import(/* webpackChunkName: "artisans" */ 'views/artisans/ArenaBalancerView') },
-  'views/CLAView': function () { return import(/* webpackChunkName: "CLAView" */ 'views/CLAView') },
-  'views/clans/ClansView': function () { return import(/* webpackChunkName: "clans" */ 'views/clans/ClansView') },
-  'views/clans/ClanDetailsView': function () { return import(/* webpackChunkName: "clans" */ 'views/clans/ClanDetailsView') },
-  'views/CommunityView': function () { return import(/* webpackChunkName: "CommunityView" */ 'views/CommunityView') },
-  'views/contribute/MainContributeView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/MainContributeView') },
-  'views/contribute/AdventurerView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/AdventurerView') },
-  'views/contribute/AmbassadorView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/AmbassadorView') },
-  'views/contribute/ArchmageView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/ArchmageView') },
-  'views/contribute/ArtisanView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/ArtisanView') },
-  'views/contribute/DiplomatView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/DiplomatView') },
-  'views/contribute/ScribeView': function () { return import(/* webpackChunkName: "contribute" */ 'views/contribute/ScribeView') },
-  'views/editor/announcement/AnnouncementSearchView': function () { return import(/* webpackChunkName: "announcementSearchView" */ 'views/editor/announcement/AnnouncementSearchView') },
-  'views/editor/announcement/AnnouncementEditView': function () { return import(/* webpackChunkName: "announcementEditView" */ 'views/editor/announcement/AnnouncementEditView') },
-  'views/editor/docs/ComponentsDocumentationView': function () { return import(/* webpackChunkName: "docs" */ 'views/editor/docs/ComponentsDocumentationView') },
-  'views/editor/docs/SystemsDocumentationView': function () { return import(/* webpackChunkName: "docs" */ 'views/editor/docs/SystemsDocumentationView') },
-  'views/editor/achievement/AchievementSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/achievement/AchievementSearchView') },
-  'views/editor/achievement/AchievementEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/achievement/AchievementEditView') },
-  'views/editor/article/ArticleSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/article/ArticleSearchView') },
-  'views/editor/article/ArticlePreviewView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/article/ArticlePreviewView') },
-  'views/editor/article/ArticleEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/article/ArticleEditView') },
-  'views/editor/level/LevelSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/level/LevelSearchView') },
-  'views/editor/level/LevelEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/level/LevelEditView') },
-  'views/editor/thang/ThangTypeSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/thang/ThangTypeSearchView') },
-  'views/editor/thang/ThangTypeEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/thang/ThangTypeEditView') },
-  'views/editor/campaign/CampaignEditorView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/campaign/CampaignEditorView') },
-  'views/editor/poll/PollSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/poll/PollSearchView') },
-  'views/editor/poll/PollEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/poll/PollEditView') },
-  'views/editor/verifier/VerifierView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/verifier/VerifierView') },
-  'views/editor/verifier/i18nVerifierView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/verifier/i18nVerifierView') },
-  'views/editor/course/CourseSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/course/CourseSearchView') },
-  'views/editor/course/CourseEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/course/CourseEditView') },
-  'views/editor/resource/ResourceSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/resource/ResourceSearchView') },
-  'views/editor/resource/ResourceEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/resource/ResourceEditView') },
-  'views/editor/podcast/PodcastSearchView': function () { return import(/* webpackChunkName: "podcastSearchView" */ 'views/editor/podcast/PodcastSearchView') },
-  'views/editor/podcast/PodcastEditView': function () { return import(/* webpackChunkName: "podcastEditView" */ 'views/editor/podcast/PodcastEditView') },
-  'views/editor/chat/ChatSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/chat/ChatSearchView') },
-  'views/editor/chat/ChatEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/chat/ChatEditView') },
-  'views/editor/ai-scenario/AIScenarioSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/ai-scenario/AIScenarioSearchView') },
-  'views/editor/ai-scenario/AIScenarioEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/ai-scenario/AIScenarioEditView') },
-  'views/editor/ai-project/AIProjectSearchView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/ai-project/AIProjectSearchView') },
-  'views/editor/ai-project/AIProjectEditView': function () { return import(/* webpackChunkName: "editor" */ 'views/editor/ai-project/AIProjectEditView') },
-  'views/i18n/I18NHomeView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NHomeView') },
-  'views/i18n/I18NEditThangTypeView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditThangTypeView') },
-  'views/i18n/I18NEditComponentView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditComponentView') },
-  'views/i18n/I18NEditLevelView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditLevelView') },
-  'views/i18n/I18NEditAchievementView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditAchievementView') },
-  'views/i18n/I18NEditCampaignView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditCampaignView') },
-  'views/i18n/I18NEditPollView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditPollView') },
-  'views/i18n/I18NEditCourseView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditCourseView') },
-  'views/i18n/I18NEditCinematicView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditCinematicView') },
-  'views/i18n/I18NEditProductView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditProductView') },
-  'views/i18n/I18NEditArticleView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditArticleView') },
-  'views/i18n/I18NEditInteractiveView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditInteractiveView') },
-  'views/i18n/I18NEditCutsceneView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditCutsceneView') },
-  'views/i18n/I18NEditResourceHubResourceView': function () { return import(/* webpackChunkName: "i18n" */ 'views/i18n/I18NEditResourceHubResourceView') },
-  'views/user/IdentifyView': function () { return import(/* webpackChunkName: "user" */ 'views/user/IdentifyView') },
-  'views/account/IsraelSignupView': function () { return import(/* webpackChunkName: "account" */ 'views/account/IsraelSignupView') },
-  'views/LegalView': function () { return import(/* webpackChunkName: "LegalView" */ 'views/LegalView') },
-  'views/play/CampaignView': function () { return import(/* webpackChunkName: "play" */ 'views/play/CampaignView') },
-  'views/ladder/LadderView': function () { return import(/* webpackChunkName: "ladder" */ 'views/ladder/LadderView') },
-  'views/ladder/MainLadderView': function () { return import(/* webpackChunkName: "ladder" */ 'views/ladder/MainLadderView') },
-  'views/ladder/MainTournamentView': function () { return import(/* webpackChunkName: "tournament" */ 'views/ladder/MainTournamentView') },
-  'views/play/level/PlayLevelView': function () { return import(/* webpackChunkName: "play" */ 'views/play/level/PlayLevelView') },
-  'views/play/level/PlayLevelVideoView': function () { return import(/* webpackChunkName: "play" */ 'views/play/level/PlayLevelVideoView') },
-  'views/play/level/PlayGameDevLevelView': function () { return import(/* webpackChunkName: "play" */ 'views/play/level/PlayGameDevLevelView') },
-  'views/play/level/PlayWebDevLevelView': function () { return import(/* webpackChunkName: "play" */ 'views/play/level/PlayWebDevLevelView') },
-  'views/play/SpectateView': function () { return import(/* webpackChunkName: "play" */ 'views/play/SpectateView') },
-  'views/PageImpact': function () { return import(/* webpackChunkName: "PageImpact" */ 'views/impact/PageImpact') },
-  'views/PagePartners': function () { return import(/* webpackChunkName: "PagePartners" */ 'views/partners/PagePartners') },
-  'views/PageAPCSPMarketing': function () { return import(/* webpackChunkName: "PageMarketing" */ 'views/apcsp/PageMarketing') },
-  'views/PageAPCSPCurriculum': function () { return import(/* webpackChunkName: "PageCurriculum" */ 'views/apcsp/PageCurriculum') },
-  'views/PremiumFeaturesView': function () { return import(/* webpackChunkName: "PremiumFeaturesView" */ 'views/PremiumFeaturesView') },
-  'views/PrivacyView': function () { return import(/* webpackChunkName: "PrivacyView" */ 'views/PrivacyView') },
-  'views/courses/CoursesView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/CoursesView') },
-  'views/courses/CoursesUpdateAccountView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/CoursesUpdateAccountView') },
-  'views/courses/ProjectGalleryView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/ProjectGalleryView') },
-  'views/courses/StudentAssessmentsView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/StudentAssessmentsView') },
-  'views/courses/ClassroomView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/ClassroomView') },
-  'views/courses/CourseDetailsView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/CourseDetailsView') },
-  'views/courses/TeacherClassesView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/TeacherClassesView') },
-  'views/courses/CourseVideosView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/CourseVideosView') },
-  'views/teachers/TeacherStudentView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/TeacherStudentView') },
-  'views/courses/TeacherClassView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/TeacherClassView') },
-  'views/courses/TeacherCoursesView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/TeacherCoursesView') },
-  'views/teachers/TeacherCourseSolutionView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/TeacherCourseSolutionView') },
-  'views/teachers/RequestQuoteView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/RequestQuoteView') },
-  'views/special_event/HoC2018View': function (){ return import(/* webpackChunkName: "HoC2018" */ 'views/special_event/HoC2018View') },
-  'views/courses/EnrollmentsView': function () { return import(/* webpackChunkName: "courses" */ 'views/courses/EnrollmentsView') },
-  'views/teachers/ResourceHubView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/ResourceHubView') },
-  'views/teachers/ApCsPrinciplesView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/ApCsPrinciplesView') },
-  'views/teachers/MarkdownResourceView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/MarkdownResourceView') },
-  'views/teachers/CreateTeacherAccountView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/CreateTeacherAccountView') },
-  'views/teachers/ConvertToTeacherAccountView': function () { return import(/* webpackChunkName: "teachers" */ 'views/teachers/ConvertToTeacherAccountView') },
-  'views/TestView': function () { return import(/* webpackChunkName: "TestView" */ 'views/TestView') },
-  'views/user/MainUserView': function () { return import(/* webpackChunkName: "user" */ 'views/user/MainUserView') },
-  'views/user/EmailVerifiedView': function () { return import(/* webpackChunkName: "user" */ 'views/user/EmailVerifiedView') },
-  'views/user/UserOptInView': function () { return import(/* webpackChunkName: "user" */ 'views/user/UserOptInView') },
-  'views/NotFoundView': function () { return import(/* webpackChunkName: "NotFoundView" */ 'views/NotFoundView') },
-  'views/teachers/RestrictedToTeachersView': function () { return import(/* webpackChunkName: "RestrictedToTeachersView" */ 'views/teachers/RestrictedToTeachersView') },
-  'views/courses/RestrictedToStudentsView': function () { return import(/* webpackChunkName: "RestrictedToStudentsView" */ 'views/courses/RestrictedToStudentsView') },
-  'views/user/CertificatesView': function () { return import(/* webpackChunkName: "CertificatesView" */ 'views/user/CertificatesView') },
-  'views/user/AnonCertificatesView': function () { return import(/* webpackChunkName: "CertificatesView" */ 'views/user/AnonCertificatesView') },
-  'views/LicensorView': function() { return import(/* webpackChunkName: "LicensorView" */ 'views/LicensorView')},
-
-  'views/landing-pages/mobile/PageMobileView': function () { return import(/* webpackChunkName: "mobile" */ 'views/landing-pages/mobile/PageMobileView') },
-
-  // All Vue app components will be loaded via a vue chunk.  Eventually we will integrate webpack loading with Vue and
-  // at that time we can remove this dynamic import and instead include SingletonAppVueComponentView in the main bundle
-  'views/core/SingletonAppVueComponentView': function () { return import(/* webpackChunkName: "vue" */ 'views/core/SingletonAppVueComponentView').then(m => m.default) },
-
-  // TODO: Why does chunk name `ozariaPlay` not work sporadically?
-  'views/ozaria/site/play/PageUnitMap': function() { return import(/*webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/play/PageUnitMap') },
-  'views/ozaria/site/avatarSelector': function() { return import(/*webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/avatar-selector/PageAvatarSelector') },
-  'views/ozaria/site/play/PagePlayLevel': function () { return import(/* webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/play/PagePlayLevel') },
-  'views/cinematic': function() { return import(/* webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/cinematic/PageCinematic') },
-  'views/cutscene': function() { return import(/* webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/cutscene/PageCutscene') },
-  'views/interactive': function() { return import(/* webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/interactive/PageInteractive') },
-  'views/introLevel': function() { return import(/* webpackChunkName: "ozariaPlay" */ 'ozaria/site/components/play/PageIntroLevel') },
-}
